♻️ refactor: ファイル名の誤字修正
Closes #127
