+++ lib/application/stamp_rally/state/upload_spot_image_result.dart
